Adds dedicated test
